Merge Two lists Sort Test Sucsess
Enter the number of elements for the first list. Then enter the elements of the list one by one using a for loop. Again, enter the elements of the second list one by one. The ‘+’ operator is then used to merge both the lists and the sort function then sorts the list in ascending order.