Finding Intersection of Two Lists
Test success.. Intersection function accepts two lists and returns the list which is the intersection of the two lists, which is,  all the common values from list 1 and 2. Hence, the set function in the intersection function accepts list and returns the list that contains the common elements in both the lists.